+++ main.go
@@ -6,6 +6,7 @@ package main
 import (
 	"gorabbit/src/cmd"
 	_ "gorabbit/src/cmd/context"
+	_ "gorabbit/src/cmd/msg"
 	_ "gorabbit/src/cmd/queues"
 	config "gorabbit/src/config"
 )