+++ src/cmd/queues/queues_delete.go
@@ -1,7 +1,7 @@
 /*
 Copyright © 2023 NAME HERE <EMAIL ADDRESS>
 */
-package queues
+package cmdqueues
 
 import (
 	"fmt"
@@ -11,23 +11,30 @@ import (
 	"github.com/spf13/cobra"
 )
 
+var queuesToDelete []string
+
 // deleteQueuesCmd represents the deleteQueues command
 var deleteQueuesCmd = &cobra.Command{
 	Use:   "delete",
-	Short: "Delete all queues on rabbitmq",
+	Short: "Delete queues from rabbitmq",
 	Run: func(cmd *cobra.Command, args []string) {
 		_, ch := api.ConnectAndGetRabbitChannel(config.CurrentContext)
-		queues := api.GetQueues(config.LocalRabbit())
-		if len(queues) <= 0 {
+
+		if len(queuesToDelete) == 0 {
+			queuesToDelete = api.GetQueues(config.CurrentContext)
+		}
+
+		if len(queuesToDelete) <= 0 {
 			fmt.Println("No queues to delete")
 		} else {
-			api.DeleteQueues(ch, queues)
+			api.DeleteQueues(ch, queuesToDelete)
 		}
 	},
 }
 
 func init() {
 	queuesCmd.AddCommand(deleteQueuesCmd)
+	deleteQueuesCmd.Flags().StringSliceVarP(&queuesToDelete, "queues", "q", []string{}, "Queues that will be deleted")
 
 	// Here you will define your flags and configuration settings.